Created a file with a color theme. Used an enum called Theme with a raw value type of String.
Created a view that summarizes the data from the DailyScrum model and displays the title, number of participants, and duration.

Created a label style for the horizontal location of the length of the battle and the clock icon. With the LabelStyle protocol, I was able to create a consistent design across my entire application by reusing the same label style across multiple views.

Applied the accessibilityAddTraits(.isHeader) modifier to the text view that displays the title of the scrum. Used the accessibility modifier to add a label describing the contents of the label in the HStack.